Level 5 start and looping platform
I've made the basic starts of level 5, namely the floor filled with a poison trap and a starting area with an elevator from the poison to the top of the starting area. I have also made a platform that moves through an array of positions. I also started to add some mathmatics so the looping platform would move in an actual circular motion, but I have commented these mechanics out until such time that we decide to use them.